refactor(web): consolidate incremental book mutation pipeline
The three incremental book mutation hooks (useIncrementalBookModify,
useIncrementalBookPatch, useIncrementalBookUpdate) each duplicated the
same document-resolution pipeline: resolve a book doc from either an id
or a document, guard against a missing document, then apply the rxdb
incremental operation. Only the operation itself differed.

Extract that shared pipeline into incrementalBookMutation and have each
hook pass its operation as a callback. Public hook APIs are unchanged.
