+++ apps/web/src/books/useIncrementalBookModify.ts
@@ -1,8 +1,7 @@
-import { getLatestDatabase } from "../rxdb/RxDbProvider"
-import { from, mergeMap, of } from "rxjs"
 import type { ModifyFunction, RxDocument } from "rxdb"
 import type { BookDocType } from "@oboku/shared"
 import { useMutation$ } from "reactjrx"
+import { incrementalBookMutation } from "./incrementalBookMutation"
 
 export const useIncrementalBookModify = () =>
   useMutation$({
@@ -13,16 +12,7 @@ export const useIncrementalBookModify = () =>
       doc: RxDocument<BookDocType> | string
       mutationFn: ModifyFunction<BookDocType>
     }) =>
-      getLatestDatabase().pipe(
-        mergeMap((db) =>
-          typeof doc === "string"
-            ? from(db.book.findOne({ selector: { _id: doc } }).exec())
-            : of(doc),
-        ),
-        mergeMap((item) => {
-          if (!item) return of(null)
-
-          return from(item.incrementalModify(mutationFn))
-        }),
+      incrementalBookMutation(doc, (item) =>
+        item.incrementalModify(mutationFn),
       ),
   })
