+++ apps/web/src/books/useIncrementalBookUpdate.ts
@@ -1,8 +1,7 @@
-import { getLatestDatabase } from "../rxdb/RxDbProvider"
-import { from, mergeMap, of } from "rxjs"
 import type { RxDocument, UpdateQuery } from "rxdb"
 import type { BookDocType } from "@oboku/shared"
 import { useMutation$ } from "reactjrx"
+import { incrementalBookMutation } from "./incrementalBookMutation"
 
 export const useIncrementalBookUpdate = () =>
   useMutation$({
@@ -13,16 +12,5 @@ export const useIncrementalBookUpdate = () =>
       doc: RxDocument<BookDocType> | string
       updateObj: UpdateQuery<BookDocType>
     }) =>
-      getLatestDatabase().pipe(
-        mergeMap((db) =>
-          typeof doc === "string"
-            ? from(db.book.findOne({ selector: { _id: doc } }).exec())
-            : of(doc),
-        ),
-        mergeMap((item) => {
-          if (!item) return of(null)
-
-          return from(item.incrementalUpdate(updateObj))
-        }),
-      ),
+      incrementalBookMutation(doc, (item) => item.incrementalUpdate(updateObj)),
   })
